+++ README.md
@@ -30,15 +30,15 @@
 
 ### 지하철역 관리 기능
 
-- [ ] 지하철역을 등록할 수 있다.
-  - [ ] 지하철역은 2글자 이상이어야 한다.
-  - [ ] 엔터키 또는 마우스 클릭으로 역을 추가할 수 있어야 한다.
-  - [ ] 지하철역의 이름은 최대 20자 이하이다.
-  - [ ] 중복된 지하철역은 추가할 수 없다.
+- [x] 지하철역을 등록할 수 있다.
+  - [x] 지하철역은 2글자 이상이어야 한다.
+  - [x] 엔터키 또는 마우스 클릭으로 역을 추가할 수 있어야 한다.
+  - [x] 지하철역의 이름은 최대 20자 이하이다.
+  - [x] 중복된 지하철역은 추가할 수 없다.
 - [ ] 지하철역의 이름을 수정할 수 있다.
   - [ ] 수정하는 UI는 modal을 이용하여 수정할 수 있게 한다.
-- [ ] 지하철역을 삭제할 수 있다.
-  - [ ] 삭제 시 confirm을 이용하여 한 번 더 유저가 확인할 수 있어야 한다.
+- [x] 지하철역을 삭제할 수 있다.
+  - [x] 삭제 시 confirm을 이용하여 한 번 더 유저가 확인할 수 있어야 한다.
   - [ ] 이미 노선에 등록된 역인 경우 삭제할 수 없다.
 
 ### 지하철 노선 관리 기능
