+++ src/js/components/LoginForm/index.ts
@@ -1,7 +1,7 @@
 import Component from "@/core/component";
-import authService from "@/service/authService";
 import view from "./view";
 import handleLink from "@/router/handleLink";
+import { authService } from "@/service";
 import { $, createElement } from "@/utils/dom";
 
 class LoginForm extends Component {
