+++ src/js/constants/base.ts
@@ -1,0 +1,5 @@
+const BASE = {
+  URL: "http://localhost:8080",
+};
+
+export default BASE;
