+++ src/js/constants/index.ts
@@ -1,3 +1,4 @@
 export { default as PAGE_TITLE } from "./pageTitle";
 export { default as PATH } from "./path";
 export { default as MESSAGE } from "./messages";
+export { default as BASE } from "./base";
