+++ src/js/service/index.ts
@@ -1,0 +1,2 @@
+export { default as authService } from "./authService";
+export { default as stationService } from "./stationService";
